# kernel: speed up IS_SUBSET_FLAGS