I investigated `IS_SUBSET_FLAGS` because callgrind shows that it is #1 in terms of instructions during GAP startup (followed by `GET_NEXT_CHAR`, `FuncWITH_HIDDEN_IMPS_FLAGS`, `strcmp` and `NextSymbol`):
```
3,617,597,065  PROGRAM TOTALS
360,902,963  src/opers.c:IS_SUBSET_FLAGS.isra.45 [/home/horn/Projekte/GAP/gap.github/gap]
299,271,263  src/io.c:GET_NEXT_CHAR [/home/horn/Projekte/GAP/gap.github/gap]
194,587,622  src/opers.c:FuncWITH_HIDDEN_IMPS_FLAGS [/home/horn/Projekte/GAP/gap.github/gap]
112,844,582  /build/glibc-Cl5G7W/glibc-2.23/string/../sysdeps/x86_64/multiarch/strcmp-sse2-unaligned.S:__strcmp_sse2_unaligned [/lib/x86_64-linux-gnu/libc-2.23.so]
 99,232,409  src/scanner.c:NextSymbol [/home/horn/Projekte/GAP/gap.github/gap]
...
```
Moreover, cachegrind indicated a huge amount of L1 read misspredictions. With this PR, it actually still has a lot, but overall timings improve.

Accessing TRUES_FLAGS requires a pointer dereference, and often leads to
a cache miss, eating up all advantage to be gained from using it; so
drop it.

To verify, I run

    echo "QUIT;" > quit.g
    perf stat -r 10 -d ./gap -A -q quit.g

on a Linux test machine, which measures the startup performance of GAP
10 times (I also verified that there was an improvement on macOS, but
less rigorously).

Before this change:

    Performance counter stats for './gap.master -A -q quit.g' (10 runs):

     760.858383      task-clock (msec)         #    0.999 CPUs utilized            ( +-  1.22% )
              7      context-switches          #    0.009 K/sec                    ( +- 12.04% )
              0      cpu-migrations            #    0.000 K/sec
            355      page-faults               #    0.467 K/sec                    ( +-  0.09% )
     2834967239      cycles                    #    3.726 GHz                      ( +-  0.38% )  (39.57%)
     1384871452      stalled-cycles-frontend   #   48.85% frontend cycles idle     ( +-  0.26% )  (39.68%)
      835157974      stalled-cycles-backend    #   29.46% backend  cycles idle     ( +-  0.26% )  (40.12%)
     3588976596      instructions              #    1.27  insns per cycle
                                               #    0.39  stalled cycles per insn  ( +-  0.42% )  (50.64%)
      834750192      branches                  # 1097.116 M/sec                    ( +-  0.32% )  (50.91%)
       15136518      branch-misses             #    1.81% of all branches          ( +-  0.57% )  (51.05%)
     1003981324      L1-dcache-loads           # 1319.538 M/sec                    ( +-  0.38% )  (49.19%)
       75716887      L1-dcache-load-misses     #    7.54% of all L1-dcache hits    ( +-  0.40% )  (20.19%)
       60158560      LLC-loads                 #   79.067 M/sec                    ( +-  1.44% )  (19.90%)
        1873602      LLC-load-misses           #    3.11% of all LL-cache hits     ( +- 10.08% )  (29.63%)

    0.761377984 seconds time elapsed                                          ( +-  1.22% )

After this change:

    Performance counter stats for './gap -A -q quit.g' (10 runs):

     718.481321      task-clock (msec)         #    0.999 CPUs utilized            ( +-  1.24% )
              8      context-switches          #    0.011 K/sec                    ( +- 11.44% )
              0      cpu-migrations            #    0.000 K/sec
            355      page-faults               #    0.494 K/sec                    ( +-  0.10% )
     2670255520      cycles                    #    3.717 GHz                      ( +-  0.11% )  (39.23%)
     1273776668      stalled-cycles-frontend   #   47.70% frontend cycles idle     ( +-  0.19% )  (39.36%)
      773319911      stalled-cycles-backend    #   28.96% backend  cycles idle     ( +-  0.45% )  (40.15%)
     3532240401      instructions              #    1.32  insns per cycle
                                               #    0.36  stalled cycles per insn  ( +-  0.10% )  (50.85%)
      809874472      branches                  # 1127.203 M/sec                    ( +-  0.18% )  (51.10%)
       12333874      branch-misses             #    1.52% of all branches          ( +-  0.29% )  (51.18%)
      960149151      L1-dcache-loads           # 1336.359 M/sec                    ( +-  0.14% )  (49.34%)
       65644690      L1-dcache-load-misses     #    6.84% of all L1-dcache hits    ( +-  0.73% )  (20.23%)
       50626980      LLC-loads                 #   70.464 M/sec                    ( +-  1.37% )  (19.81%)
        3477462      LLC-load-misses           #    6.87% of all LL-cache hits     ( +-  1.50% )  (29.43%)

    0.719034384 seconds time elapsed                                          ( +-  1.24% )

So we go from ~0.76 seconds to ~0.72 seconds startup time on average on this machine.

I also checked the testinstall testsuite, but less thoroughly, using

    perf stat -r 5 -d ./gap -A -q tst/testinstall.g

Before: ~56.408 (+- 0.15%) seconds time elapsed
After: ~56.280 (+- 0.15%) seconds time elapsed

Here it provides not much of an advantage; but it also doesn't cause a slow down.


I suspect we could speed flags up quite a bit more by making use of the fact that they are usually very sparse. E.g. I looked at the flag lists used for implications, and while they were up to ~2200 entries long or beyond, they had at most 90 bits set to true; but on average only 7.82, median 3.
```
gap> printStats:=function(vals)
>     Print("min ", Minimum(vals), " - avg ", Int(100*Average(vals))/100.,
>           " - med ", Median(vals), " - max ", Maximum(vals), "\n");
> end;;
gap> filters:=Concatenation(Flat(IMPLICATIONS_SIMPLE),Flat(IMPLICATIONS_COMPOSED));;
gap> printStats(List(filters, SIZE_FLAGS));
min 1 - avg 7.82 - med 3 - max 90
gap> printStats(List(filters, LEN_FLAGS));
min 1 - avg 1024.97 - med 1017 - max 2210
gap> printStats(List(filters, SIZE_OBJ));
min 40 - avg 164.15 - med 160 - max 312
```
Also note how the average flag object takes 160 bytes, so multiple flags together will quickly exceed the L1 cache.  Thus, perhaps with a more clever encoding (even a simple RLE scheme), we could gain a lot more, simply by virtue of it leading to much smaller flags objects which better fit into the cache. 


BTW, I also tried reversing the order we loop over memory in `IS_SUBSET_FLAGS`, thinking that it might help the CPU in predicting the memory access better, but this actually caused a slowdown. I have not yet investigate why, but I think part of the reason is that the flags we compare often are identical in the first few blocks.