@ChrisJefferson 2018-05-03
I seem to remember I previously flipped the order, for exactly the reason you describe :)

I didn't realise the first bit of code isn't actually required.