@ChrisJefferson 2018-05-03
I dug up some code i had about this function. It is called
11,959,225 times during startup. Of those, 97% of the time we detect they are not subsets when we compare the last pair of blocks, with this PR. Therefore if you try adjusting, I'd still do that check first. Of the remaing 3%, about half pass and half fail.

Intersting thing I just noticed, if I add an assert just after 'if(len1 < len2)' with 'assert(ptr2[len2-1] != 0)', it passes -- do we promise not to put empty blocks on the end of BLOCKS_FLAGS? if so, the whole first branch can be reduced to 'return 0'. I suspect this won't greatly increase the speed as we are exiting almost straight away anyway.