@fingolfin 2018-05-03
@ChrisJefferson We do not make any such promise, and it doesn't hold:
```
gap> f:=SUB_FLAGS(f,f);
<flag list>
gap> LEN_FLAGS(f);
112
gap> TRUES_FLAGS(f);
[  ]
```
But of course we could change the code to do so -- given that creating filters is done far less often than testing them, this might pay off... And actually, I think `SUB_FLAGS` is the only function where this can happen, and changing it to resize the flags bag to minimal size shouldn't be hard (and can be done quite efficiently).