@fingolfin 2018-05-03
@stevelinton we do *NOT* guarantee "equal iff identical" for flag lists, and indeed, this is frequently violated.

Storing a list of 16 bit flag numbers is one of various possible alternate data formats, yes. For the worst case with 90 true flags, that would amount to 180 bytes (plus header), but of course the smaller ones are indeed much, much better.

If we always store the flags as short as possible, we can also get rid of the `LEN_FLAGS` field (also, on the GAP level, there is only one call to it in `lib/oper1.g:605` which can and should be changed to use `TRUES_FLAGS` instead -- after which we could get rid of both `FuncLEN_FLAGS` and `FuncELM_FLAGS`). We could then also get rid of the `TRUES_FLAGS` field, and instead return a fresh plist each time `TRUES_FLAGS` is called -- the only two performance critical places calling it are in `RunImmediateMethods`; but the first one of those is making  a shallow copy of it anyway, the second appends it to another list -- so we could remove `FuncTRUES_FLAGS` and replace it by a `FuncAPPEND_TRUES_FLAGS`.


Of course there are other possible formats: e.g. various variations of run length encoding, such as: store runs as signed bytes: for each byte `n`, if `n>=0` then this indicates `n+1` zero bits; otherwise, `-n` one bits. Since very often two adjacent bits are set (for properties), this can sometimes save a bit; but given the extreme sparsity of our flags in practice, I am not sure the extra complication would be worth it.

Just out of curiosity, I experimented a bit to confirm this; while I don't think this line of research is going to be fruitful, here's my test code snippet anyway, which takes a trues list `t` and "compresses" it into ranges:
```
gap> t:=TRUES_FLAGS(IMPLICATIONS_COMPOSED[1][1]);
[ 15, 35, 56, 69, 70, 88, 89, 90, 99, 102, 106, 110, 114, 140, 141, 235 ]
gap> r:=[];; i:=1;; while i <= Length(t) do n:=t[i]; m:=n; while i<Length(t) and m+1=t[i+1] do i:=i+1;m:=m+1; od; Add(r, [n..m]); i:=i+1; od; r;
[ [ 15 ], [ 35 ], [ 56 ], [ 69, 70 ], [ 88 .. 90 ], [ 99 ], [ 102 ], [ 106 ], [ 110 ], [ 114 ], [ 140, 141 ], [ 235 ] ]
gap> Length(t); Length(r);
16
12
```