@frankluebeck 2018-05-03
I tried the patch on some machines. It is difficult to judge if there is a speedup (one is mainly measuring the strategy of the system to change the speed of CPU cores). But there is certainly no obvious slow down and therefore the simpler code is welcome. 

For a possible improvement I would also try to store arrays of 16-bit unsigned shorts. But how much memory is needed by all current bit lists in a typical GAP session? There is probably not much to save. And when `IS_SUBSET_FLAGS` returns `false` in most cases, this is probably often detected very early in the bit list comparison. So, it is not immediately clear, if a change will actually lead to a performance gain that justifies the efforts.